[NTFS] Fix MSVC warnings in mkntfs.c and dispatch.c
- mkntfs.c: Cast cluster_size to ULONGLONG in NTFS_ALIGN_UP to avoid
  C4319 (zero extending ULONG to ULONGLONG via bitwise NOT).
- dispatch.c: Remove redundant DPRINT that left Stack unused when
  the macro is compiled out on amd64, triggering C4189 under /we4189.

+++ sdk/lib/fslib/ntfslib/mkntfs.c
@@ -1232,7 +1232,7 @@ int mkntfs_format(const MKNTFS_IO *io, const MKNTFS_PARAMS *params)
     s.logfile_size = total_bytes / 400;
     if (s.logfile_size < 2 * 1024 * 1024) s.logfile_size = 2 * 1024 * 1024;
     if (s.logfile_size > 64 * 1024 * 1024) s.logfile_size = 64 * 1024 * 1024;
-    s.logfile_size = NTFS_ALIGN_UP(s.logfile_size, s.cluster_size);
+    s.logfile_size = NTFS_ALIGN_UP(s.logfile_size, (ULONGLONG)s.cluster_size);
     s.logfile_clusters = s.logfile_size / s.cluster_size;
     s.logfile_lcn = next_lcn;
     bitmap_set(s.lcn_bitmap, s.logfile_lcn, s.logfile_clusters);
